% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_mob_impact.R
\name{active_mob_impact}
\alias{active_mob_impact}
\title{Compute the economic impact associated with a shift from passive to active mobility}
\usage{
active_mob_impact(
  scenario,
  met_phy_act = 7,
  rr_phy_bike = 0.899,
  rr_phy_walk = 0.883,
  rr_phy_ebike = 0.899,
  exp_level_base = 11.25,
  max_met_walk = 31,
  max_met_bike = 51,
  max_met_ebike = 51,
  rr_pm25 = 1.08,
  voly = 70000,
  vsl = 3600000,
  voly_vls_ref_yr = 2016,
  air_pollution = TRUE,
  injuries = TRUE,
  detail = FALSE
)
}
\arguments{
\item{scenario}{A \code{data.frame} describing a mobility-change scenario defined with \code{\link[=scenario_builder]{scenario_builder()}}.}

\item{met_phy_act}{A numeric value. The physical effort associated with moderate physical activity (expressed in Mets).
This parameter is used quantify the overall weekly physical activity of new active mobility adopters
performing sports in their free-time. Default is \code{7} MET.}

\item{rr_phy_bike}{A numeric value. Relative risk reduction associated with cycling. The default value is \code{0.10} and is taken from
the HEAT manual.}

\item{rr_phy_walk}{A numeric value. Relative risk reduction associated with walking. The default value is \code{0.11} and is taken from
the HEAT manual.}

\item{rr_phy_ebike}{A numeric value. Relative risk reduction associated with riding an e-bike and is assumed to be the same of that
of bikes. The default value is \code{0.10}.}

\item{exp_level_base}{A numeric value. The minimum level of weekly physical activity (expressed in METs) above
which an individual experiences health benefits. The default value is \code{11.25}.}

\item{max_met_walk}{A numeric value. The level of physical activity exerted walking above which no additional
health benefits can be obtained. The default value, from the HEAT manual, is \code{31}.}

\item{max_met_bike}{A numeric value. The level of physical activity exerted cycling above which no additional
health benefits can be obtained. The default value, from the HEAT manual, is \code{51}.}

\item{max_met_ebike}{A numeric value. The level of physical activity exerted riding an e-bike above which no additional
health benefits can be obtained. The default value is \code{51}, assumed equal to that of bikes.}

\item{rr_pm25}{A numeric value. The relative risk (RR) of death associated with a 10 ug/m3 incrase in background
PM2.5 concentrations. The default value of \code{1.08} is taken from Chen & Hoek (increase in mortality
risk due to a 10 ug/m3 increase in PM2.5).}

\item{voly}{A numeric value. The Value of a Life Year to be used in the economic valuation. The
default value is \code{70000} euro 2016 et refers to the EU, taken from the 2019 Handbook of the external cost of transport.}

\item{vsl}{A numeric value. The value of a statistical life, to be used in the economic valuation. The
default value is \code{3.6} million euro (2016) and refers to the EU. Taken from the 2019 Handbook of the external cost of transport.}

\item{voly_vls_ref_yr}{A numeric value. The year to which the \code{voly} and the \code{vsl} refer. Default is \code{2016}.}

\item{air_pollution}{A logical value. If set to \code{TRUE} the negative health effects associated with increased
air pollutant intake are considered. Default value is set to \code{TRUE}.}

\item{injuries}{A logical value. If set to \code{TRUE} the negative health effects associated with increased
road injury risk are considered. Default value is set to \code{TRUE}.}

\item{detail}{A logical value. If set to \code{TRUE} a detailed version of the output is provided, with specific
estimates for each age group, physical activity level and year. Default is set to \code{FALSE}, which return
aggregate values over the whole length of the scenario.}
}
\value{
a tibble with the outcome of the economic evaluation: avoided deaths and their economic value estimated
with both the VOLY and the VSL. Values per km travelled are provided if \code{detail} is set to \code{FALSE}.
}
\description{
This function quantifies the health impacts (mortality only) associated with a shift from passive
to active mobility and provides their economic valuation using both the VOLY and the VSL.
Overall health impacts are given by the positive impact associated to increased
physical activity and the negative impacts of increased air pollutant intake and road
injury risk.
}
\examples{
mode_change_shares <- c(0.2, 0.15, 0.1, 0.05)

scenario_builder("Palermo", "private_car_driver", "bike", mode_change_distance = mode_change_shares, comm_days = 4, max_km = 10)

active_mob_impact(scenario)

active_mob_impact(scenario, met_phy_act = 6, detail = TRUE)
}
