% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fat_risk.R
\docType{data}
\name{fat_risk}
\alias{fat_risk}
\title{Fatality risk estimates for Italy}
\format{
A data frame with two variables
\describe{
\item{mode}{Active mobility mode}
\item{fat_rate}{Fatality rate (deaths/km travelled)}
}
}
\usage{
fat_risk
}
\description{
A dataset with fatality risk estimates for Italy. Data are provided for walking,
cyling and riding an e-bike.
}
\keyword{datasets}
