% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob_share_demand.R
\docType{data}
\name{mob_share_demand}
\alias{mob_share_demand}
\title{Data on total mobility shares and demand in Italy}
\format{
A data frame with three variables.
\describe{
\item{year}{Year}
\item{mobility_share}{Share of total population commuting}
\item{mobility_demand}{Trips per day}
}
}
\source{
\url{https://www.isfort.it/}
}
\usage{
mob_share_demand
}
\description{
A dataset containing data - taken from Isfort reports - on mobility shares and
mobility data referring to Italy.
}
\keyword{datasets}
