% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_act.R
\docType{data}
\name{phy_act}
\alias{phy_act}
\title{Physical activity levels}
\format{
A tibble with three variables.
\describe{
\item{duration}{Duration of physical activity, minutes/week}
\item{age}{age group}
\item{phy_act_share}{Share of the population spending time in health-enhancing physical activity}
}
}
\source{
\url{https://ec.europa.eu/eurostat/data/database}
}
\usage{
phy_act
}
\description{
A dataset containing data on the physical activity levels among Italian citiziens.
Data are about the time spent on health-enhancing (non-work-related) aerobic physical activity,
by age group. Data refer to Italy as a whole and to the year 2019.
Age group in the original dataset are different from those used in this package. An adaptation was made.
}
\keyword{datasets}
