% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_act_impact.R
\name{phy_act_impact}
\alias{phy_act_impact}
\title{Quantifies the health benefits associated with increased physical activity linked to
active mobility and compute their economic value.}
\usage{
phy_act_impact(
  scenario,
  met_phy_act = 7,
  rr_0_bike = 0.1,
  rr_0_walk = 0.11,
  rr_0_ebike = 0.1,
  exp_level_base = 11.25,
  max_met_walk = 31,
  max_met_bike = 51,
  max_met_ebike = 51,
  base_yr = FALSE,
  voly = 70000,
  vsl = 3600000,
  voly_vls_ref_yr = 2016,
  detail = FALSE
)
}
\arguments{
\item{scenario}{A \code{data.frame} describing a mobility-change scenario defined with \code{\link[=scenario_builder]{scenario_builder()}}.}

\item{met_phy_act}{A numeric value. The physical effort associated with moderate physical activity (expressed in Mets).
This parameter is used quantify the overall weekly physical activity of new active mobility adopters
performing sports in their free-time. Default is \code{7} MET.}

\item{rr_0_bike}{A numeric value. Relative risk reduction associated with cycling. The default value is \code{0.10} and is taken from
the HEAT manual.}

\item{rr_0_walk}{A numeric value. Relative risk reduction associated with walking. The default value is \code{0.11} and is taken from
the HEAT manual.}

\item{rr_0_ebike}{A numeric value. Relative risk reduction associated with riding an e-bike and is assumed to be the same of that
of bikes. The default value is \code{0.10}.}

\item{exp_level_base}{A numeric value. The minimum level of weekly physical activity (expressed in METs) above
which an individual experiences health benefits. The default value is \code{11.25}.}

\item{max_met_walk}{A numeric value. The level of physical activity exerted walking above which no additional
health benefits can be obtained. The default value, from the HEAT manual, is \code{31}.}

\item{max_met_bike}{A numeric value. The level of physical activity exerted cycling above which no additional
health benefits can be obtained. The default value, from the HEAT manual, is \code{51}.}

\item{max_met_ebike}{A numeric value. The level of physical activity exerted riding an e-bike above which no additional
health benefits can be obtained. The default value is \code{51}, assumed equal to that of bikes.}

\item{base_yr}{A logical value. The reference year of the analysis. The year to which all economic values refer.
The default value is \code{FALSE}, which means that the reference year is taken from the \code{scenario} dataframe.}

\item{voly}{A numeric value. The Value of a Life Year to be used in the economic valuation. The
default value is \code{70000} euro 2016 et refers to the EU, taken from the 2019 Handbook of the external cost of transport.}

\item{vsl}{A numeric value. The value of a statistical life, to be used in the economic valuation. The
default value is \code{3.6} million euro (2016) and refers to the EU. Taken from the 2019 Handbook of the external cost of transport.}

\item{voly_vls_ref_yr}{A numeric value. The year to which the \code{voly} and the \code{vsl} refer. Default is \code{2016}.}

\item{detail}{A logical value. If set to \code{TRUE} a detailed version of the output is provided, with specific
estimates for each age group, physical activity level and year. Default is set to \code{FALSE}, which return
aggregate values over the whole length of the scenario.}
}
\value{
a tibble with the outcome of the economic evaluation: avoided deaths and their economic value estimated
with both the VOLY and the VSL. Values per km travelled are provided if \code{detail} is set to \code{FALSE}.
}
\description{
The function takes as a main input a mobility-change scenario defined with \code{\link[=scenario_builder]{scenario_builder()}}.
Other parameters all have a default value but can be modified by the user.
Only effects on mortality are considered.
}
\examples{
mode_change_shares <- c(0.2, 0.15, 0.1, 0.05)

scenario_builder("Palermo", "private_car_driver", "bike", mode_change_distance = mode_change_shares, comm_days = 4, max_km = 10)

phy_act_impact(scenario)
phy_act_impact(scenario, met_phy_act = 6, detail = TRUE)
}
