% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_imp.R
\name{phy_imp}
\alias{phy_imp}
\title{Computing variables of interest to assess the health impact associated with increased physical activity}
\usage{
phy_imp(
  data,
  met_phy_act,
  max_met_walk,
  rr_phy,
  max_met_bike,
  max_met_ebike,
  mode_to,
  exp_level_base
)
}
\arguments{
\item{data}{A data frame obtained with \code{scenario_builder}.}

\item{met_phy_act}{A numeric value. The physical effort associated with moderate physical activity (expressed in Mets).
This parameter is used quantify the overall weekly physical activity of new active mobility adopters
performing sports in their free-time.}

\item{max_met_walk}{A numeric value. The level of physical activity exerted walking above which no additional
health benefits can be obtained.}

\item{rr_phy}{A numeric value. Relative risk reduction associated with the active mobility mode assesed}

\item{max_met_bike}{A numeric value. The level of physical activity exerted cycling above which no additional
health benefits can be obtained.}

\item{max_met_ebike}{A numeric value. The level of physical activity exerted e-cycling above which no additional
health benefits can be obtained.}

\item{mode_to}{A string. The active mode to which the shift occurs.}

\item{exp_level_base}{A numeric value. The minimum level of weekly physical activity (expressed in METs) above
which an individual experiences health benefits.}
}
\value{
A data frame similar to the input data frame but including variables of interest to compute the health impact
increased physical activity
}
\description{
It is an internal function, used by \code{active_mob_impact}.
}
\keyword{internal}
