% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_25_urban_ita.R
\docType{data}
\name{pm_25_urban_ita}
\alias{pm_25_urban_ita}
\title{Mean PM2.5 background concentrations in Italian cities}
\format{
A tibble with eight variables.
\describe{
\item{name}{City name}
\item{bkg_conc}{PM2.5 mean background concentration, ug/m~3~}
\item{code}{Municipality ISTAT code}
\item{region}{Region name}
\item{province}{Province name}
\item{nuts1}{NUTS1 code}
\item{nuts2}{NUTS2 code}
\item{nuts3}{NUTS3 code}
}
}
\source{
\url{https://www.who.int/publications/m/item/who-air-quality-database-2022}
}
\usage{
pm_25_urban_ita
}
\description{
A dataset of mean annual background concentrations of PM2.5 in some Italian
cities. Data are from the WHO Air quality Database 2022 and refer to the period
2010 - 2019. The mean is computed over this period. Not all Italian municipalities
are in the database.
}
\keyword{datasets}
