% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollution_imp.R
\name{pollution_imp}
\alias{pollution_imp}
\title{Computing variables of interest to assess the health impact associated with increased air pollutant intake}
\usage{
pollution_imp(data, rr_pm25, nuts3, bkg_conc, speeds, mode_from, mode_to)
}
\arguments{
\item{data}{A data frame obtained with \code{scenario_builder}.}

\item{rr_pm25}{A numeric value. The relative risk (RR) of death associated with a 10 ug/m3 incrase in background
PM2.5 concentrations. The default value of \code{1.08} is taken from Chen & Hoek (increase in mortality
risk due to a 10 ug/m3 increase in PM2.5).}
}
\value{
A data frame similar to the input data frame but including variables of interest to compute the health impact
increased air pollutant intake
}
\description{
It is an internal function, used by \code{active_mob_impact}.
}
\keyword{internal}
