% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollution_impact.R
\name{pollution_impact}
\alias{pollution_impact}
\title{Quantify the (negative) health impact of increased exposure to air pollutants linked to
active mobility with respect to passive mobility}
\usage{
pollution_impact(
  scenario,
  rr_25 = 1.08,
  voly = 70000,
  vsl = 3600000,
  voly_vls_ref_yr = 2016,
  detail = FALSE
)
}
\arguments{
\item{scenario}{A \code{data.frame} describing a mobility-change scenario defined with \code{\link[=scenario_builder]{scenario_builder()}}.}

\item{rr_25}{A numeric value. The relative risk (RR) of death associated with a 10 ug/m3 incrase in background
PM2.5 concentrations. The default value of \code{1.08} is taken from Chen & Hoek (increase in mortality
risk due to a 10 ug/m3 increase in PM2.5)}

\item{voly}{A numeric value. The Value of a Life Year to be used in the economic valuation. The
default value is \code{70000} euro 2016 et refers to the EU, taken from the 2019 Handbook of the external cost of transport.}

\item{vsl}{A numeric value. The value of a statistical life, to be used in the economic valuation. The
default value is \code{3.6} million euro (2016) and refers to the EU. Taken from the 2019 Handbook of the external cost of transport.}

\item{voly_vls_ref_yr}{A numeric value. The year to which the \code{voly} and the \code{vsl} refer. Default is \code{2016}.}

\item{detail}{A logical value. If set to \code{TRUE} a detailed version of the output is provided, with specific
estimates for each age group, physical activity level and year. Default is set to \code{FALSE}, which return
aggregate values over the whole length of the scenario.}
}
\value{
a tibble with the outcome of the economic evaluation: additional deaths and their economic value estimated
with both the VOLY and the VSL. Values per km travelled are provided if \code{detail} is set to \code{FALSE}.
}
\description{
The function takes as a main input a mobility-change scenario defined with \code{\link[=scenario_builder]{scenario_builder()}}.
Other parameters all have a default value but can be modified by the user.
Only effects on mortality are considered.
}
\examples{
mode_change_shares <- c(0.2, 0.15, 0.1, 0.05)

scenario_builder("Palermo", "private_car_driver", "bike", mode_change_distance = mode_change_shares, comm_days = 4, max_km = 10)

phy_act_impact(scenario)
pollution_impact(scenario)
}
