% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_evolution.R
\name{pop_evolution}
\alias{pop_evolution}
\title{Calculate the evolution of a population given a mortality rate}
\usage{
pop_evolution(pop, death_rate)
}
\arguments{
\item{pop}{A numeric value. The initial number of individuals.}

\item{death_rate}{A numeric vector. A vector of annual mortality rates. A vector
is used to allow the user to apply different mortality rates for different years.}
}
\value{
A vector of population values of the same length of \code{death_rate}.
}
\description{
This function calculates the year-by-year evolution of a population given the
yearly mortality rate. It is useful to compute avoided and additional deaths
in mobility change scenarios. The input for year \emph{n} is the output of year \emph{n-1}.
}
\examples{
actmobilityr:::pop_evolution(500, seq(from = 0.001, to = 0.0006, by = - 0.0001))
}
\keyword{internal}
