% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_ita.R
\docType{data}
\name{pop_ita}
\alias{pop_ita}
\title{Total residing population in Italy}
\format{
A tibble with two variables.
\describe{
\item{year}{Year}
\item{pop}{Residing population on January 1st}
}
}
\source{
\url{https://ec.europa.eu/eurostat/data/database}
}
\usage{
pop_ita
}
\description{
A data frame containing data on the total population of Italy. Data are from Eurostat
and are used to quantify changes in road injury risk.
}
\keyword{datasets}
