% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_builder.R
\name{scenario_builder}
\alias{scenario_builder}
\title{Define a commuting mode change scenario}
\usage{
scenario_builder(
  city,
  mode_from,
  mode_to,
  mode_change_share,
  comm_days,
  max_km
)
}
\arguments{
\item{city}{A \code{character} string.: the city for which the scenario needs to be built. The user must
insert a valid name (starting with a capital letter) of any of the Italian municipalities.
For the full list of available municipalities see \code{mun_codes}. No multiple names allowed.}

\item{mode_from}{A \code{character} string.: the transportation mode used by commuters before the
commuting mode change. For the full list of modes available see private_car_driver, private_car_passenger, motorbike_scooter, bike, other, walk, urban_bus, extra_urban_bus, train, school_company_bus, tram, and subway}

\item{mode_to}{A \code{character} string : the active transportation mode to which individuals shift. The
user cn choose from: \code{walk}, \code{bike} and \code{ebike}. No multiple names allowed.}

\item{mode_change_share}{A numeric value: the share of total commuters shifting from passive to active mobility. Value
must be between \code{0} and \code{1}. Alternatively to a single value, also a vector of values can be provided,
to specify a different value for each age group in \code{demo_data}. Age groups range from
\code{20-24} to \code{60-64} years. Therefore the first value of the vector is the share of the \code{20-24} years  of age
group, while the last refers to the \code{60-64} group. Each value must be between \code{0} and \code{1}. Vector length
must be the same of that of the age groups considered.}

\item{comm_days}{A numeric value. Indicates the number of commuting days per week. Values must be between \code{1}
and \code{5}.}

\item{max_km}{A numeri value. The distance threshold (km) below which a given share of commuters (defined with
\code{mode_change_share}) shift from passive to active mobility. This distance is a one-way distance.
For example, when \code{max_km = 5}, a given share of all commuting travels below (or equal) to 5 km one-way
are performed with active modes of travel instead than passive.}

\item{age_specific}{A logical value: use \code{FALSE} to assume that the share of total commuters.
who shift to active mobility is the same across age groups. Use \code{TRUE} to provide age-specific values
(see \code{mode_change_share}).}
}
\value{
A data.frame with the commuting mode change scenario, to be used to estimate health benefits.
}
\description{
This function creates a tibble with all relevant information for a mobility change
scenario for any Italian city and for several transportation means.
}
\examples{
scenario_builder("Palermo", "private_car_driver", "bike", 0.2, 4, 5)

## creating a vector of mode change shares

shares <- c(0.4, 0.3, 0.3, 0.3, 0.2, 0.2, 0.2, 0.1, 0.1)

scenario_builder("Palermo", "private_car_driver", "bike", shares, 4, 5)

}
