% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speeds_met.R
\docType{data}
\name{speeds_met}
\alias{speeds_met}
\title{Dataset of average speeds and energy expenditure of active mobility modes}
\format{
A tibble with three variables
\describe{
\item{mode}{Active mobility mode}
\item{speed}{Average speed, km/h}
\item{met}{Energy expenditure, METs}
}
}
\usage{
speeds_met
}
\description{
A dataset containing information on average speeds (km/h) and energy expenditure
expressed in METs.
}
\keyword{datasets}
