% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_speeds.R
\docType{data}
\name{transport_speeds}
\alias{transport_speeds}
\title{Average speeds of transportation means}
\format{
A tibble with 8 variables:
\describe{
\item{mean_of_transp}{Transportation mean}
\item{speed_kmh}{Average speed, km/h}
}
}
\source{
{https://ec.europa.eu/eurostat/data/database}
}
\usage{
transport_speeds
}
\description{
A dataset of average speeds of transportation means expressed in km/h
}
\keyword{datasets}
